% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socoastal-package.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\alias{ibcso_500}
\alias{gebco_500}
\alias{scar_0}
\title{Bathymetric isolines}
\description{
Geographic line objects, in simple features format (\code{sf} package).
}
\details{
ibcso_500 is the -500m contour from IBCSO
gebco_500 is the -500m contour from GEBCO 2014
scar_0 is the coastline from CCAMLR, the CCAMLR Coastline based on SCAR Coastline data from \code{CCAMLRGIS}
See "data-raw/bathymetry.R" for extraction details.
}
