% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socoastal-package.R
\docType{package}
\name{socoastal-package}
\alias{socoastal}
\alias{socoastal-package}
\title{socoastal: Explore Antarctica Coastal Regionalization}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
